requirements.txt
streamlit==1.32.0
torch>=2.0.0
ftfy
regex
tqdm
Pillow
git+https://github.com/openai/CLIP.git

